app/requirements.txt
Django==3.1
gunicorn==20.0.4
psycopg2-binary==2.8.5
djangorestframework==3.11.1
django-filter==2.3.0
django-cors-headers==3.5.0
